pyproject.toml
[tool.poetry]
name = "trip-planner"
version = "0.1.0"
description = "AI-based trip planner"
authors = ["aryan-jain <technicallyaryan@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.32.2"
langchain = "^0.1.14"
langchain-openai = "^0.1.1"
python-dotenv = "^1.0.1"
streamlit-pydantic = "^0.6.0"
pydantic = "~1.10.11"
st-pages = "^0.4.5"
streamlit-extras = "^0.4.2"
alembic = "^1.13.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
st-supabase-connection = "^2.0.0"
streamlit-javascript = "^0.1.5"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
tzlocal = "^5.2"

[tool.poetry.scripts]
trip-planner = "trip_planner.Home:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
